app_backend/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-JWT-Extended==4.6.0 
waitress